Project File Structure modified and Working Borders are Added :)
