+++ src/test/resources/features/VehiclesPage.feature
@@ -1,6 +1,8 @@
+
 @wip
 Feature: As a user, I should be able to arrange vehicle table data under Fleet-Vehicle page
 
+  # ESALKAN xFleet Project Features STARTS HERE
   # 1- User can arrange rows/vehicle numbers to be displayed by clicking on 'View Per Page' button under Fleet-Vehicles module
   Scenario Outline: View Per Page Button Options Feature
     Given the user is on the login page
@@ -83,3 +85,6 @@ Feature: As a user, I should be able to arrange vehicle table data under Fleet-V
 #      | username        | password    | title     |
 #      | salesmanager120 | UserUser123 | Dashboard |
 #      | storemanager85  | UserUser123 | Dashboard |
+
+
+  # ESALKAN xFleet Project Features ENDS HERE
